Shift welcomeMsg & exitMsg from main() to class variables of Scribbles
Move welcomeMsg & exitMsg as constant variables under Scribbles as I want the job
of main() function to control the flow and interaction of Scribbles tentatively.
Therefore, there is no need for main() to hold these 2 variables.

Side effect: The constants are accessible outside of the class due to it being 'public',
may require changes in the future.