Update logic flow in main() to support different task types
With the addition of different types of tasks, there is a need to handle each tasks
separately with new commands to identify the tasks that users want to add. In addition,
depending on the task that users want to add, parameters are required for these tasks.

Side note: Need to work on handling errors and exceptions to ensure parameters are
enforced for certain tasks as well