Update logic flow in main() to handle different errors
Handle different exceptions and throw the appropriate error message.

Error handling is important to test for cases where users might not input
a command that follows the user guide

Side note: Exception handling current is extremely messy and there is definitely
a need to refactor on. Very bug prone probably as well.